+++ src/dragen_align_pa/jobs/manage_dragen_mlr.py
@@ -1,17 +1,19 @@
 import json
+import os
 import subprocess
-import time
+from collections.abc import Callable
+from functools import partial
+from typing import Any
 
 import cpg_utils
 from cpg_flow.targets import Cohort
-from cpg_utils.config import config_retrieve, get_driver_image, try_get_ar_guid
+from cpg_utils.config import config_retrieve, get_driver_image
 from cpg_utils.hail_batch import get_batch
 from hailtop.batch.job import PythonJob
 from loguru import logger
 
-from dragen_align_pa.constants import ICA_CLI_SETUP
-from dragen_align_pa.jobs import cancel_ica_pipeline_run, monitor_dragen_pipeline
-from dragen_align_pa.utils import delete_pipeline_id_file
+from dragen_align_pa.constants import BUCKET_NAME, ICA_CLI_SETUP
+from dragen_align_pa.jobs.ica_pipeline_manager import manage_ica_pipeline_loop
 
 
 def _initalise_mlr_job(cohort: Cohort) -> PythonJob:
@@ -23,65 +25,177 @@ def _initalise_mlr_job(cohort: Cohort) -> PythonJob:
     return mlr_job
 
 
+def _run_command(
+    command: str | list[str],
+    capture_output: bool = False,
+    shell: bool = False,
+) -> subprocess.CompletedProcess:
+    """
+    Runs a subprocess command with robust error logging.
+    """
+    executable = '/bin/bash' if shell else None
+    cmd_str = command if isinstance(command, str) else ' '.join(command)
+
+    try:
+        logger.info(f'Running command: {cmd_str}')
+        return subprocess.run(
+            command,
+            check=True,
+            text=True,
+            capture_output=capture_output,
+            shell=shell,
+            executable=executable,
+        )
+    except subprocess.CalledProcessError as e:
+        logger.error(f'Command failed with return code {e.returncode}: {cmd_str}')
+        if e.stdout:
+            logger.error(f'STDOUT: {e.stdout.strip()}')
+        if e.stderr:
+            logger.error(f'STDERR: {e.stderr.strip()}')
+        # Re-raise as a generic exception to fail the job
+        raise ValueError('A subprocess command failed. See logs.') from e
+
+
+def _find_ica_file_path(parent_folder: str, file_name: str) -> str:
+    """
+    Finds a file in ICA and returns its full `details.path`.
+    """
+    command = [
+        'icav2',
+        'projectdata',
+        'list',
+        '--parent-folder',
+        parent_folder,
+        '--data-type',
+        'FILE',
+        '--file-name',
+        file_name,
+        '--match-mode',
+        'EXACT',
+        '-o',
+        'json',
+    ]
+    result = _run_command(command, capture_output=True)
+    try:
+        data = json.loads(result.stdout)
+        if not data.get('items'):
+            raise ValueError(f'No file found with name "{file_name}" in folder "{parent_folder}"')
+
+        file_path = data['items'][0].get('details', {}).get('path')
+        if not file_path:
+            raise ValueError(f'File "{file_name}" found, but it has no "details.path" in API response.')
+
+        logger.info(f'Found {file_name} at path: {file_path}')
+        return file_path
+
+    except json.JSONDecodeError:
+        logger.error(f'Failed to decode JSON from icav2 list command: {result.stdout}')
+        raise
+    except (ValueError, IndexError) as e:
+        logger.error(f'Error parsing icav2 list output for {file_name}: {e}')
+        raise
+
+
 def _submit_mlr_run(
     pipeline_id_arguid_path: cpg_utils.Path,
-    bucket: str,
     ica_analysis_output_folder: str,
     sg_name: str,
     mlr_project: str,
     mlr_config_json: str,
     mlr_hash_table: str,
     output_prefix: str,
-    is_bioheart: bool,
 ) -> str:
-    # All of bioheart was run through alignment and genotyping before the implementation of the ar_guid in ICA
-    # Therefore the paths need to be corrected for that
-    if is_bioheart:
-        with pipeline_id_arguid_path.open() as pid_arguid_fhandle:
-            pipeline_id: str = pid_arguid_fhandle.readline().strip()
-            ar_guid: str = ''
-    else:
-        with pipeline_id_arguid_path.open() as pid_arguid_fhandle:
-            data: dict[str, str] = json.load(pid_arguid_fhandle)
-            pipeline_id = data['pipeline_id']
-            ar_guid = f'_{data["ar_guid"]}_'
+    """
+    Submits the DRAGEN MLR pipeline by running individual CLI commands
+    and parsing the JSON output file.
+    """
+    with pipeline_id_arguid_path.open() as pid_arguid_fhandle:
+        data: dict[str, str] = json.load(pid_arguid_fhandle)
+        pipeline_id = data['pipeline_id']
+        ar_guid = f'_{data["ar_guid"]}_'
 
-    mlr_analysis_command: str = f"""
-        # General authentication
-        {ICA_CLI_SETUP}
-        cram_path=$(icav2 projectdata list --parent-folder /{bucket}/{ica_analysis_output_folder}/{sg_name}/{sg_name}{ar_guid}-{pipeline_id}/{sg_name}/ --data-type FILE --file-name {sg_name}.cram --match-mode EXACT -o json | jq -r '.items[].details.path')
-        gvcf_path=$(icav2 projectdata list --parent-folder /{bucket}/{ica_analysis_output_folder}/{sg_name}/{sg_name}{ar_guid}-{pipeline_id}/{sg_name}/ --data-type FILE --file-name {sg_name}.hard-filtered.gvcf.gz --match-mode EXACT -o json | jq -r '.items[].details.path')
-        cram="ica://OurDNA-DRAGEN-378${{cram_path}}"
-        gvcf="ica://OurDNA-DRAGEN-378${{gvcf_path}}"
-
-        icav2 projects enter {mlr_project}
-
-        # Download the mlr config JSON
-        icav2 projectdata download {mlr_config_json} $BATCH_TMPDIR/mlr_config.json --exclude-source-path > /dev/null 2>&1
-
-        popgen-cli dragen-mlr submit \
-        --input-project-config-file-path $BATCH_TMPDIR/mlr_config.json \
-        --output-analysis-json-folder-path {sg_name} \
-        --run-id {sg_name}-mlr \
-        --sample-id {sg_name} \
-        --input-ht-folder-url {mlr_hash_table} \
-        --output-folder-url {output_prefix}/{sg_name}{ar_guid}-{pipeline_id}/{sg_name} \
-        --input-align-file-url ${{cram}} \
-        --input-gvcf-file-url ${{gvcf}} \
-        --analysis-instance-tier {config_retrieve(['ica', 'mlr', 'analysis_instance_tier'])} > /dev/null 2>&1
-
-        cat {sg_name}/sample-{sg_name}-run-{sg_name}-mlr.json | jq -r ".id"
-    """  # noqa: E501
-    mlr_analysis_id: str = (
-        subprocess.run(mlr_analysis_command, shell=True, capture_output=True, check=False).stdout.decode().strip()
+    batch_tmpdir = os.environ.get('BATCH_TMPDIR', '/batch')
+    local_config_path = os.path.join(batch_tmpdir, 'mlr_config.json')
+    ica_base_folder = (
+        f'/{BUCKET_NAME}/{ica_analysis_output_folder}/{sg_name}/{sg_name}{ar_guid}-{pipeline_id}/{sg_name}/'
     )
 
-    return mlr_analysis_id
+    try:
+        # --- 1. Authenticate ---
+        # shell=True is required for the multi-line ICA_CLI_SETUP script
+        _run_command(ICA_CLI_SETUP, shell=True)
+
+        # --- 2. Find input file paths ---
+        cram_path = _find_ica_file_path(ica_base_folder, f'{sg_name}.cram')
+        gvcf_path = _find_ica_file_path(ica_base_folder, f'{sg_name}.hard-filtered.gvcf.gz')
+
+        cram_url = f'ica://OurDNA-DRAGEN-378${cram_path}'
+        gvcf_url = f'ica://OurDNA-DRAGEN-378${gvcf_path}'
+
+        # --- 3. Set ICA project context ---
+        _run_command(['icav2', 'projects', 'enter', mlr_project])
+
+        # --- 4. Download MLR config JSON ---
+        _run_command(
+            [
+                'icav2',
+                'projectdata',
+                'download',
+                mlr_config_json,
+                local_config_path,
+                '--exclude-source-path',
+            ]
+        )
+
+        # --- 5. Build and run the popgen-cli command ---
+        # This is built as a list to avoid shell=True
+        submit_command: list[str] = [
+            'popgen-cli',
+            'dragen-mlr',
+            'submit',
+            '--input-project-config-file-path',
+            local_config_path,
+            '--output-analysis-json-folder-path',
+            sg_name,
+            '--run-id',
+            f'{sg_name}-mlr',
+            '--sample-id',
+            sg_name,
+            '--input-ht-folder-url',
+            mlr_hash_table,
+            '--output-folder-url',
+            f'{output_prefix}/{sg_name}{ar_guid}-{pipeline_id}/{sg_name}',
+            '--input-align-file-url',
+            cram_url,
+            '--input-gvcf-file-url',
+            gvcf_url,
+            '--analysis-instance-tier',
+            config_retrieve(['ica', 'mlr', 'analysis_instance_tier']),
+        ]
+        _run_command(submit_command)
+
+        # --- 6. Read the pipeline ID from the output JSON ---
+        output_json_path = os.path.join(sg_name, f'sample-{sg_name}-run-{sg_name}-mlr.json')
+        if not os.path.exists(output_json_path):
+            raise FileNotFoundError(f'popgen-cli did not produce expected output file: {output_json_path}')
+
+        with open(output_json_path) as f:
+            submission_data: dict[str, Any] = json.load(f)
+
+        mlr_analysis_id = submission_data.get('id')
+        if not mlr_analysis_id:
+            raise ValueError(f'Submission output file "{output_json_path}" is missing the "id" key.')
+
+        logger.info(f'MLR pipeline ID for {sg_name} is {mlr_analysis_id}')
+        return mlr_analysis_id
+
+    except (subprocess.CalledProcessError, FileNotFoundError, ValueError, json.JSONDecodeError) as e:
+        logger.error(f'Failed to submit MLR pipeline for {sg_name}: {e}')
+        raise
 
 
 def run_mlr(
     cohort: Cohort,
-    bucket: str,
     pipeline_id_arguid_path_dict: dict[str, cpg_utils.Path],
     outputs: dict[str, cpg_utils.Path],
 ) -> PythonJob:
@@ -89,161 +203,54 @@ def run_mlr(
 
     job.call(
         _run,
-        cohort,
-        bucket,
-        pipeline_id_arguid_path_dict,
+        cohort=cohort,
+        pipeline_id_arguid_path_dict=pipeline_id_arguid_path_dict,
         outputs=outputs,
     )
 
     return job
 
 
-def _run(  # noqa: PLR0915
+def _run(
     cohort: Cohort,
-    bucket: str,
     pipeline_id_arguid_path_dict: dict[str, cpg_utils.Path],
     outputs: dict[str, cpg_utils.Path],
 ) -> None:
-    logger.info('Starting MLR processing and monitoring')
-    # Error sink needs to be configured here
-    logger.add(sink='tmp_errors.log', format='{time} - {level} - {message}', level='ERROR')
-
+    """
+    Calls the generic pipeline manager with settings for the MLR pipeline.
+    """
     ica_analysis_output_folder: str = config_retrieve(['ica', 'data_prep', 'output_folder'])
     mlr_project: str = config_retrieve(['ica', 'projects', 'dragen_mlr'])
     dragen_align_project: str = config_retrieve(['ica', 'projects', 'dragen_align'])
     mlr_config_json: str = config_retrieve(['ica', 'mlr', 'config_json'])
     mlr_hash_table: str = config_retrieve(['ica', 'mlr', 'mlr_hash_table'])
 
-    is_bioheart: bool = 'bioheart' in cohort.dataset.name
     logger.info(f'Dataset name is: {cohort.dataset.name}')
-    logger.info(f'Is Bioheart = {"bioheart" in cohort.dataset.name}')
 
-    running_pipelines: list[str] = []
-    cancelled_pipelines: list[str] = []
-    failed_pipelines: list[str] = []
-    # If a sequencing group is in this list, we can skip checking its status
-    completed_pipelines: list[str] = []
+    def _create_submit_callable(sg_name: str) -> Callable[[], str]:
+        """Creates a zero-argument callable for pipeline submission."""
+        output_prefix: str = f'ica://{dragen_align_project}/{BUCKET_NAME}/{config_retrieve(["ica", "data_prep", "output_folder"])}/{sg_name}'
 
-    while (len(completed_pipelines) + len(cancelled_pipelines) + len(failed_pipelines)) < len(
-        cohort.get_sequencing_groups()
-    ):
-        for sequencing_group in cohort.get_sequencing_groups():
-            sg_name: str = sequencing_group.name
-            mlr_pipeline_success_file: cpg_utils.Path = outputs[f'{sg_name}_mlr_success']
-            mlr_pipeline_id_file: cpg_utils.Path = outputs[f'{sg_name}_mlr_pipeline_id']
-            with pipeline_id_arguid_path_dict[f'{sg_name}_pipeline_id_and_arguid'].open() as arguid_fh:
-                if is_bioheart:
-                    ar_guid: str = try_get_ar_guid()
-                else:
-                    ar_guid = json.load(arguid_fh)['ar_guid']
+        return partial(
+            _submit_mlr_run,
+            pipeline_id_arguid_path=pipeline_id_arguid_path_dict[f'{sg_name}_pipeline_id_and_arguid'],
+            ica_analysis_output_folder=ica_analysis_output_folder,
+            sg_name=sg_name,
+            mlr_project=mlr_project,
+            mlr_config_json=mlr_config_json,
+            mlr_hash_table=mlr_hash_table,
+            output_prefix=output_prefix,
+        )
 
-            output_prefix: str = f'ica://{dragen_align_project}/{bucket}/{config_retrieve(["ica", "data_prep", "output_folder"])}/{sg_name}'  # noqa: E501
-
-            # In case of Hail Batch crashes, find previous completed runs so we can skip trying to monitor them
-            if mlr_pipeline_success_file.exists() and sg_name not in completed_pipelines:
-                completed_pipelines.append(sg_name)
-                continue
-
-            if any(
-                [
-                    sg_name in completed_pipelines,
-                    sg_name in cancelled_pipelines,
-                    sg_name in failed_pipelines,
-                ]
-            ):
-                continue
-
-            # Get an MLR analysis ID for each sequencing group, if it exists.
-            mlr_analysis_id: str = ''
-            if mlr_pipeline_id_file_exists := mlr_pipeline_id_file.exists():
-                with mlr_pipeline_id_file.open('r') as mlr_pid_fhandle:
-                    mlr_analysis_id = json.load(mlr_pid_fhandle)['pipeline_id']
-
-            # Handle cancelling a pipeline first
-            # Cancel a running job in ICA
-            if config_retrieve(key=['ica', 'management', 'cancel_cohort_run'], default=False) and mlr_analysis_id:
-                logger.info(f'Cancelling pipeline run: {mlr_analysis_id} for sequencing group {sg_name}')
-                cancel_ica_pipeline_run.run(ica_pipeline_id=mlr_analysis_id, is_mlr=True)
-                delete_pipeline_id_file(pipeline_id_file=str(mlr_pipeline_id_file))
-            else:
-                # If a pipeline ID file doesn't exist we have to submit a new run, regardless of other settings
-                if not mlr_pipeline_id_file_exists:
-                    logger.info(f'Submitting MLR pipeline run for {sg_name}')
-                    mlr_analysis_id = _submit_mlr_run(
-                        pipeline_id_arguid_path=pipeline_id_arguid_path_dict[f'{sg_name}_pipeline_id_and_arguid'],
-                        bucket=bucket,
-                        ica_analysis_output_folder=ica_analysis_output_folder,
-                        sg_name=sg_name,
-                        mlr_project=mlr_project,
-                        mlr_config_json=mlr_config_json,
-                        mlr_hash_table=mlr_hash_table,
-                        output_prefix=output_prefix,
-                        is_bioheart=is_bioheart,
-                    )
-                    logger.info(f'MLR pipeline ID for {sg_name} is {mlr_analysis_id}')
-                    with mlr_pipeline_id_file.open('w') as mlr_fhandle:
-                        mlr_fhandle.write(json.dumps({'pipeline_id': mlr_analysis_id, 'ar_guid': ar_guid}))
-
-                mlr_pipeline_status: str = monitor_dragen_pipeline.run(ica_pipeline_id=mlr_analysis_id, is_mlr=True)
-
-                if mlr_pipeline_status == 'INPROGRESS':
-                    running_pipelines.append(sg_name)
-
-                elif mlr_pipeline_status == 'SUCCEEDED':
-                    logger.info(f'Pipeline run {mlr_analysis_id} has succeeded for {sg_name}')
-                    completed_pipelines.append(sg_name)
-                    # Testing fix
-                    if sg_name in running_pipelines:
-                        running_pipelines.remove(sg_name)
-                    # Write the success to GCP
-                    with mlr_pipeline_success_file.open('w') as success_file:
-                        success_file.write(
-                            f'ICA pipeline {mlr_analysis_id} has succeeded for sequencing group {sg_name}.'
-                        )
-
-                elif mlr_pipeline_status in ['ABORTING', 'ABORTED']:
-                    logger.info(f'The pipeline run {mlr_analysis_id} has been cancelled for sample {sg_name}.')
-                    cancelled_pipelines.append(sg_name)
-                    if sg_name in running_pipelines:
-                        running_pipelines.remove(sg_name)
-                    delete_pipeline_id_file(pipeline_id_file=str(mlr_pipeline_id_file))
-
-                elif mlr_pipeline_status in ['FAILED', 'FAILEDFINAL']:
-                    # Log failed ICA pipeline to a file somewhere
-                    if sg_name in running_pipelines:
-                        running_pipelines.remove(sg_name)
-                    failed_pipelines.append(sg_name)
-                    delete_pipeline_id_file(pipeline_id_file=str(mlr_pipeline_id_file))
-                    logger.error(
-                        f'The pipeline {mlr_analysis_id} has failed, deleting pipeline ID file {sg_name}_pipeline_id'
-                    )
-
-        # If some pipelines have been cancelled, abort this pipeline
-        # This code will only trigger if a 'cancel pipeline' run is submitted before this master pipeline run is
-        # cancelled in Hail Batch
-        if cancelled_pipelines:
-            # If there are both cancelled and failed pipelines,
-            # write the failed pipelines to the error log before exiting.
-            if failed_pipelines:
-                with open('tmp_errors.log') as tmp_log_handle:
-                    lines: list[str] = tmp_log_handle.readlines()
-                    with outputs[f'{cohort.name}_errors'].open('a') as gcp_error_log_file:
-                        gcp_error_log_file.write('\n'.join(lines))
-            raise Exception(f'The following MLR pipelines have been cancelled: {" ".join(cancelled_pipelines)}')
-
-        # If more than 5% of pipelines are failing, exit now so we can investigate
-        if failed_pipelines and float(len(failed_pipelines)) / float(len(cohort.get_sequencing_groups())) > 0.05:  # noqa: PLR2004
-            raise Exception(
-                f'More than 5% of MLR pipelines have failed. Failing pipelines: {" ".join(failed_pipelines)}'
-            )
-        # Catch just in case everything is finished on the first pass through the loop
-        if (len(completed_pipelines) + len(cancelled_pipelines) + len(failed_pipelines)) == len(
-            cohort.get_sequencing_groups()
-        ):
-            break
-        # Wait 5 1/2 minutes before checking again (average run time 20 minutes)
-        time.sleep(330)
-    with open('tmp_errors.log') as tmp_log_handle:
-        lines = tmp_log_handle.readlines()
-        with outputs[f'{cohort.name}_mlr_errors'].open('w') as gcp_error_log_file:
-            gcp_error_log_file.write('\n'.join(lines))
+    manage_ica_pipeline_loop(
+        cohort=cohort,
+        outputs=outputs,
+        pipeline_name='MLR',
+        is_mlr_pipeline=True,
+        success_file_key_template='{sg_name}_mlr_success',
+        pipeline_id_file_key_template='{sg_name}_mlr_pipeline_id',
+        error_log_key=f'{cohort.name}_mlr_errors',
+        submit_function_factory=_create_submit_callable,
+        allow_retry=False,
+        sleep_time_seconds=330,
+    )
